refactor: routes components
