+++ packages/client/src/routes/404.tsx
