+++ packages/client/src/routes/Login.tsx
@@ -1,20 +1,12 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
+import styled from "styled-components";
 
-import {
-  Container,
-  Title,
-  Box,
-  Register,
-  Button,
-} from "./styles";
-
-import axios from '../../axios'
-
+import axios from '../axios'
 
 // components
-import Input from "../../components/Input";
-import Alert from "../../components/Alert";
+import Input from "../components/Input";
+import Alert from "../components/Alert";
 import Snackbar from '@mui/material/Snackbar';
 
 const Login: React.FC = (props) => {
@@ -88,3 +80,48 @@ const Login: React.FC = (props) => {
 };
 
 export default Login;
+
+export const Container = styled.div`
+  display: flex;
+  height: 100vh;
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
+
+export const Title = styled.h2`
+  color: black;
+`;
+
+export const Box = styled.div`
+  width: 50%;
+  height: 100%;
+  margin-top: 20px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+`;
+
+export const Register = styled.button`
+  color: #363636;
+  background-color: transparent;
+  border: none;
+  font-size: 14px;
+  align-self: flex-end;
+  font-weight: bold;
+  cursor: pointer;
+`;
+
+export const Button = styled.button`
+  background-color: #D7FF61;
+  height: 50px;
+  border: none;
+  width: 100%;
+  margin-top: 60px;
+  border-radius: 5px;
+  font-weight: bold;
+  font-size: 16px;
+  cursor: pointer;
+`
