+++ packages/client/src/routes/PrivateRouter.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Navigate, RouteProps } from "react-router";
-import auth from "../../auth";
+import auth from "../auth";
 
 const PrivateRouter = ({ children }: RouteProps) => {
   if (!auth()) {
