--- packages/client/src/routes/Transactions/styles.tsx
@@ -1,16 +1,0 @@
-import styled from "styled-components";
-
-export const Container = styled.div`
-  width: 100%;
-`;
-
-export const Button = styled.button`
-  background-color: #D7FF61;
-  height: 50px;
-  border: none;
-  width: 100%;
-  border-radius: 5px;
-  font-weight: bold;
-  font-size: 16px;
-  cursor: pointer;
-`
